+++ test/packages/monorepo-nested-root/npm/functions/another/src/exclude-me.js
@@ -1,0 +1,3 @@
+"use strict";
+
+module.exports = "I should be excluded by serverless.yml patterns";